Attempt 3 - Updated weekly Club Report
Expands the weekly Club Report.  If Options>End week report descriptive
details is enabled (EWD = 1), then it displays a weekly summary report
for each assigned Club slave like the current Penthouse Report does.
Notable exception, the DJ (if assigned) does NOT display her "SA serves
the public" summary as it seemed redundant when compared to current Club
Report summary; a lot of the same info in both.  If Option> EWD is
disabled (EWD = 0) then it simply tacks on a list of all Club slaves at
the bottom of the current report.