Update saRivalries.tw
More of a code reduction this time, but maybe a bit of a gain in speed too. Again I removed most of the for j loops. I could of made it into a big for loop but the smaller one should be faster, since there can only be on rival no need to loop over all the code, just need to find the one.