Bugfixes-Optimized brothelReport.tw
Bugfixes include now reads the right results for relationships/rivalries for the Madam due to it using an un-updated $Madam var after running the "Sa's" ($slaves[]). Also added in use of canDoVaginal(), canDoAnal(), and canPenetrate() for the Madam. Madams rival will be kept now and not wiped out when the variable gets cleaned up at end of week, as it is also now saved with her slaves[] counterpart.

Also fixed a typo for the $oralTotal += 10 it was analTotal += 10.
Changed some variables over to _temps; _brothelBonuses, _profits, and _madamFetish as they are only used/needed in this passage.