Bugfixes-Optimizations dairyReport.tw
Fixes issues with the $Milkmaid var and her $slaves[] var being out of sync. Also used canPenetrate() for her. An I fixed her penetration counts to be the same # of times as for a slave (+10 for each)
Changed Line 420 <<if ($Milkmaid.muscles > 1)>> to be <<if ($Milkmaid.muscles > 30)>> As that seems to be a left over from the old system.

Made some of variables as _vars since they are only used in this passage _anusesStretched, _birthers, _births, _cumWeek, _femCumWeek, _milkmaidFetish, _milkmaidImpregnated, _milkWeek. The _birthers, _births vars don't seem to do anything anymore (beyond disabling the related texts for them), but I left them in just in case it's for use or something.