Bugfix-Opt clinicReport.tw
Not much changed again mainly just keeping the $Nurse var in sync. Again I separated the Nurse code from the patients.

Changed more vars into _temp vars;  _idleBonus, _nurseFetish, and _restedSlaves.
I deleted the $nurseUsedCure as it's not used anywhere, and $spaSlaves-- var as I assumed it was supposed to actually be $clinicSlaves--, but later on noticed that the _restedSlaves var did the same as $clinicSlaves--, but later on in the code so I took it out as well.