Update storyInit.tw
Forgot to make changes to this file when I converted $SQMilk, and $SQMilkSale to _temp vars.

Also noticed that the $Vars are back in at least the servantsquartersReport.tw I guess the merge conflicts didn't do it right. I'll add the corrected version also here, and check the other files as well just in case.