Update assign.tw
Code reductions.

Of NOTE: don't put goto's in case statements as it'll skip them for some reason, and continue off at the end of the switch. Hmm or maybe it doesn't like goto'ing "strings" anymore as the if version does the same I did more or less having to skip all the other code.