Update spaReport.tw
Fixed an exploit. Attendant now can only cure a mindbroken slave or fix a flaw, not both the same week and not more than on flaw a week as it was, seems it was intended for only one slave's flaw a week. Also I noticed that if you have an Attendant with 100 a mindbroken slave is almost guaranteed to be cured. if (random(1,100) > 90-$Attendant.devotion) (90 - 100 = -10) Even if I abs()'ed it it still would be a 90% success rate, while fixing a flaw only has a 60% chance of success.
Also fixed the removing of slaves from the spa, and ththe count once removed.