Automated multiple organ and limb surgery (#3)
Adds the ability to have all completed organs and limbs automatically implanted.

Follows the same rules as manual implantation.
Will not discard unimplantable organs or unattachable limbs.
Gives a single screen of text with all the slave reactions to their surgeries.
Note: the current descriptions do not give the slave names, and can be confusing. surgeryDegradation.tw should be updated to lead with the slave's name.

Also: Added functionality to automatically discard organs and limbs for slaves you don't have.