Huge performance boost by not saving 500KB of duplication information
On every click of a link the game is saved.  But this includes all of the $heroSlaves
of which there are 127 taking up an average of 4KB each!

The result is having to json'ify, then compress and encode 500KB of data on every single link press,
making it all very slow.

This is a simple fix of not cloning the whole activeSlave each time, but only the relevant details
and then doing the full clone only when you actually need to create a new slave.

The result is a huge performance boost.