Add <<textinput>> macro
This is like a <<textarea>> but also like a link: on every change the
user makes, the code that is enclosed is run.

So, for example,

    <<textinput "$somevar" "">>
        <<replace "#someid">>
            $somevar
        <</replace>>
    <</textinput>>

will mirror everything that the user writes to the element #someid.

I'm mostly adding this for the coming changes but there may be other
parts of the game that could use such a macro.