Rework the Rule Assistant activation
The only actual change is from the current "activation" mechanism and
its variables to using an AST-like value, that is then evaluated to
check if the rule applies.

All the other changes are to support that.  There are minimal interface
changes.  The UI stays the same, the underlying data structures that it
manipulates are quite different though.  In js/rulesAssistant.tw are
functions that manipulate the AST to make any changes that the UI asks
for.

You may ask: why?  The reason is the following commit...