Change the way rules are applied
Instead of iterating through every slave and then every rule, seeing if
it applies and actually applying it, making the changes it requires, we
merge the relevant rules into one and then apply that.

The only problem has to do with setting assignments or assigning to
facilities.  In particular, with the exclude special slaves option.
Right now, we "revert" back to looping over the rules, which results in
some silly messages sometimes, like "She is now working in the brothel.
She is now resting.  She is now a servant."  That's not new behaviour,
but it still be nice if it could be "solved".