Various RA fixes
Previously, if you wrote a negative number in a custom rule and that
custom rule was "simple" in every other way, the UI would still refuse to
display it, since it was only checking for "(number)" and a negative
number has a "-" surrounding it.  So write a pass that turns every
negative number in a "(number)".

Also, if you had written a custom rule with a syntax error and the
turned to an "activation" UI, upon saving it would still complain about
the syntax error.  That's because the saving code was only checking for
the existance of "_customCondition", and that was set when you clicked
on "Custom".  So now, unset "_customCondition" when clicking any
"activation", but save its value so that all your changes are still
there if you come back.  Of course, you leave the RA page everything's
still gone.

And various other small stuff.  It also tries to fix import/export.