Sorted assignments and transfers
Visually set apart two categories of assignment: Slave improvement and
slave workplace.  Hopefully, this will make it easier to find at a
glance what assigment the master is looking for.

Also ensured that roles are in the same order between assignments and
improvements.

This means they are in a different order than RA for now though.

Takes up a bit more vertical space, and there's no precedent I can find
for splitting one variable accross multiple lines like this, but I can't
think of a better way to set it up so far.