Sort order in slave summary
Now consistant with both RA and slave interact.

Does explicitly saying "Improvement/Work" here seem to confuse things?
I shortened it a bit, but it could be dropped completely.

As in other places, had to change Fuckdoll check to check every
assignment individually, instead of once for a block.  I don't know if
it slows things down, but it allows the order to be changed freely.