added the missing ability to have sex with fucktoy
Previously fucktoys (toychest) with no preference can't be sexed up from the same menu, which doesn't make sense since if you change the preferred hole to anything you suddenly can. I believe this is because the code was incomplete. I added the ability to have sex with the fucktoys from the main menu.