Bug fix for the Sweatshop earnings
Previously the earnings for the Sweatshop is bugged. The way it's coded, fully staffing a Sweatshop will give the player less money and the presence of a single Sweatshop makes all helots earn more, even if there are more helots than the Sweatshop can use. This bugfix should make the Sweatshop give the correct amount of extra money.