slave.recruiter bugfix
Consistency bugfix. Previously when the activeSlave has the .recruiter value of older sister, sometimes the code treats the activeSlave as the older sister and sometimes it treats the recruited sibling as the older sister. Now the code is consistent. When activeSlave.recruiter has the value of older sister, the activeSlave is the older slave and the recruited slave is the younger sister. Vice versa for activeSlave.recruiter is young sister.