Usual donkey layout work
newSlave options are now broken up into sections so favorite options are
easier to pinpoint.  Option at top based on the slave's case, a section
that focuses on changing the slave, and a section that focuses on sexy
times.

Some options now share a line, to make the list easier to read
vertially.