Increased attrxx/attrxy changes in longtermeffects
Checked an old version and sure enough, these values date back to when attraction was a -2 to +2 scale. Since we're on a 100-point scale instead of a 5-point scale now, that makes these values a twentieth as powerful as they were back then. Cranked everything up between two and fivefold - still significantly slower, but reasonable.