Fix for LegendaryCow and LegendaryBalls prestige
The LegendaryCow and LegendaryBalls prestige text in the brothel and few other locations will make reference to slaves producing cum and milk even if they no longer do so. The proposed change would catch if the slave no longer fits the criteria that gave them the prestige and draw attention to it.

If certain future societies are in effect, having the slave no longer fit the criteria is a boon, and fitting the criteria is no longer considered helpful.

This is my first proposed change, I hope I didn't commit any faux-pas or otherwise mess things up!