Additional fixes (#1)
* Fixed spelling error

* Fixed grammatical error

* Added missing space

* Fixed visual inconsistency when changing rules

* Updated old face values

* Fixed check in rival origin

I believe the inequalities were reversed. Previously a player with $seeDicks=0 would only get the options with dicks, and vice versa for players with $seeDicks=100. Additionally, if the player had a different value for $seeDicks, there was the possibility of generating an empty array which would force the "addict" result. If this was intended you can simply change "<=" to "<" to revert it.

* Modified conditions for RE devotees and RE rebels

Previously the condition for these two events was dependent on $slaves being sorted by devotion in descending order. This removes that dependency.

* Compatibility update

Now compatible with the changed conditions for the event selection in randomNonIndividualEvent.tw

* Compatibility update

Now compatible with the changed conditions for the event selection in randomNonIndividualEvent.tw

* Fixed visual inconsistency

If a player had enough more penthouse slaves than beds, number of penthouse beds would be red, but if the same were true for luxury rooms the number would be white.

If this was intended feel free to ignore/revert.

* Updated clickability of some options

Implemented this suggestion from reddit:
> you really need to keep all options clickable in all fields, always. This is already the case for the most part but not in all fields. For example in the speech rules, if you have restrictive chosen and then choose permissive the restrictive option is no longer availlable, so if you change your mind you can no longer choose restrictive unless you confirm the changes and then get back in, which is frustrating.