Mostly Textbox Additions
Added more textboxes to ease backwards compatibility. Given that this menu is mostly supposed to be for bug tests, it should also make it easier to check weird values for error handling or testing new values for old stats.

Also added ability to modify lactationAdaption, Age Implant, DNA damage and a few other odds and ends and touch-ups.