I blue myself
Bug: Caught missing "extremely pale" from darkening options.

Bodypaint/Dye: Rather than make dye/paint a seperate variable, I'm adding this here.

Drawbacks would include the loss of the "underlying" skin description (Would it be worth defining an $activeSlave.origSkin?) but tigers go well with sharp nails, Camo with certain outfist and bodyguards (Cammy?)

Also there about a bajillion different bodypaint ideas, but this covers the basics I think.

Darkening/lightening options disappear when skin is dyed, and are replaced by an option for a new skin: "natural."  "Natural" acts as "dye removed" and allows skin to lightened or darkened to taste.

This required reworking "darken" which used to act as a sort of catch all for some reason.