Industrial dairy inflation
Inflation is always a weird fetish, and I don't want this to be too immersion breaking.  I've tried to indulge fantasy as well as nodding to the slave's safety with this magical 'gel,' but that might be too much Handwavium for Free Cities.

I've also tried to leave the master a bit free in his fancy as he watches, leaving players to fill in the image as they please.

I've checked it with a few slaves now, caught some bugs, seems good.

That means I'll commit and find something wrong immediately.  D: