Major pass condensing piercing options in bodyModification
The idea here is to condense things a bit from the stock 13 lines, to a flexible system that goes down to 3 if slave is unpierced.

Minimally, seperating the act of piercing from describing piercing will make it easier to adjust and bug fix code.

This also allows for a description section, (and seperates it from options), rather than a clinical list.  Although sometimes lists are nice.  Now that it's seperated, it would be trivial to restore the list format if that is prefered, and simply hide unpierced lines.

In particular, heavy piercings need to be more desciptive and so get a bit clunky, with, commas.

Minor: added heavy corset piercing.  Light now has ribbons, heavy chains.