WIP help needed
I am banging my head against the wall. When lines 1506-1534 are present, `<<if $activeSlave.brand != 0>>` seems to break off from it's `else` partner and I get redtext.  I can get a single `if` to work, but no more.

I tried it with `if  /if` and more recently with `if elseif`.

I'm sure it's stupid obvious, but I can't see it.