Add links to rivals and relations, color them appropriately.
I hope it's ok to add this to the stylesheet, it's the only way I can get the color of the link to respect the color of the passage.

This will be useful elsewhere I'm sure, but it's possible that someone else was assuming links would always be blue, and encased one in @@.  If that's the case, it's just a matter of correctly escaping the link itself from the palette choice.