Somehow "bovine" was also gone
My fault.  Not sure how I missed having it.

That exposed another problem in the way things are handled in bovine, though, which the new if conditional checks.  Basically, bovine's shoulder slot describes arms and upper/lower back, and then the ACTUAL arms and back slots are empty.  This produces white lines in the way bodyMod was calling widget descriptions.

This is a kludge on a kludge, though.  It might be better to simply (boringly) describe her cow tats in the same way as the others, slot by slot.