Tweak "Refreshment delivery event" plural refreshments
It used to read something like: "carrying a selection of cigar and the necessary implements on a tray. "

I tried to make it "cigars" but there would be mouse/mouses edge cases.

With small tweaks, the plurals now refer to a collection/selection of a single thing, as in:

`carrying a selection of cigar brands and the necessary implements on a tray. `