Major Bugfix for shoulder and hip surgery
Looking at legs, I think this is the way it's supposed to work:

You can make a change ONCE. Then, you need to upgrade the surgery and you can expand and contract with advanced surgery as many times as you please.

Before this fix, it was very easy to get "locked out" of being able to do surgeries on shoulders and hips after the first.

Minor: just check once for indenture restrictions.