Tweaks: Warn about skill loss in the correct place and only when needed
Short of my pending split of `VaginaDescription` into a seperate widget for dicks, this changes two things:

Hides surgery warnings for the vagina/dick section if no vagina is present.  (Ain't no dickSkill.)

Second, ensure that the skill description comes last in a mouth description, so that when a skills warning is added, it's in the right place.

Minor: fix missing skill modifier in surgeryDegradation