Tabs to Main
This adds tabs for each role to the Main screen.  Currently, it's very simple.

It does not "save" which tab player was on during page refresh.  Frustrating.  I'm a bit over my head with what they did here with the `defaultOpen` tag

Ideally, I'd like to be able to take a tally of every slave roll, perhaps show that number of slaves on the tab, and hide the tab if not in use.  Not sure where we tally.

Maybe there is a more compressed way to do this in code.

Later I'll also replace the "tabs" I added to the spa and such with these shiny ones.  I'd like to thank whoever made these?

Minor: indent changes to slave interact is all that happens below line 542, I was trying to learn what the code did.