Variant: Slightly differect way to do home with tabs
Here I've tried to bring the "will be your/ Manage X" lines closer to the actual slave in question in an overview tab.  It could stand to be prettier, but it has some real benefits.  Less scrolling.  Less duplication of information, especially on a home screen.  Helps to connect this particular slave with the big gams to this particular role.  Really nice with vector art.

Better? Worse?

Tested keyboard shortcuts for HG and such, they all still work.