Eliminate .Index from facility arrays
I guess it was there for performance reasons, but it seems kinda
pointless.  The current solution should be about as fast and it's
certainly simpler.

Basicaly, we just create an Object that associates slave IDs with their
indexes and its saved in $slaveIndices.  This is created right now
everytime <<SlaveSort>> is called, however it doesn't need to be, it
could be just done at the start of the reports and reused by all of
them (since its main use is in the reports).  Something for the future I
guess.